requirements.txt
absl-py==0.8.1
astor==0.8.0
gast==0.3.2
grpcio==1.24.1
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.15.4
Pillow==6.2.0
protobuf==3.10.0
PyYAML==5.1.2
scipy==1.3.1
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
Werkzeug==0.16.0
